backend/requirements.txt
fastapi==0.110.0
starlette==0.37.2
pycryptodome